Move JpegFile to N3Base.
This file was maintained by the developers, since it adds a layer of
encryption of the jpeg file, which is used by KSC images.
