+++ src/engine/N3Base/JpegFile.cpp
