+++ src/engine/N3Base/JpegFile.h
